+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - 更新 ESLint 配置 | Update ESLint configuration
 - 更新捐助信息 | Update donation information
+- 调整 API `/api/system/getConf` | Adjust API `/api/system/getConf`
 
 ## v0.3.14 / 2024-11-20